Add GoldenLayoutFactory class and use it in DashboardRenderer classes.
